+++ rollup.config.js
@@ -1,11 +1,14 @@
+import nodeResolve from 'rollup-plugin-node-resolve'
 const pkg = require('./package.json')
 
 export default {
   entry: 'index.js',
   moduleName: 'globalMercator',
   sourceMap: true,
-  legacy: true,
+  plugins: [nodeResolve({jsnext: true})],
   targets: [
-    { dest: pkg['main'], format: 'umd' }
+    { dest: pkg['main'], format: 'cjs' },
+    { dest: pkg['browser'], format: 'umd' },
+    { dest: pkg['jsnext:main'], format: 'es' }
   ]
 }