+++ index.js
@@ -1,5 +1,5 @@
-const originShift = 2 * Math.PI * 6378137 / 2.0
-function initialResolution (tileSize = 256) { return 2 * Math.PI * 6378137 / tileSize }
+var originShift = 2 * Math.PI * 6378137 / 2.0
+function initialResolution (tileSize) { return 2 * Math.PI * 6378137 / tileSize || 256 }
 
 /**
  * Hash tile for unique id key
@@ -7,11 +7,13 @@ function initialResolution (tileSize = 256) { return 2 * Math.PI * 6378137 / til
  * @param {Tile} tile [x, y, z]
  * @returns {number} hash
  * @example
- * const id = hash([312, 480, 4])
+ * var id = hash([312, 480, 4])
  * //=5728
  */
 export function hash (tile) {
-  const [x, y, z] = tile
+  var x = tile[0]
+  var y = tile[1]
+  var z = tile[2]
   return (1 << z) * ((1 << z) + x) + y
 }
 
@@ -21,13 +23,16 @@ export function hash (tile) {
  * @param {BBox} bbox - [west, south, east, north] coordinates
  * @return {LngLat} center
  * @example
- * const center = bboxToCenter([90, -45, 85, -50])
+ * var center = bboxToCenter([90, -45, 85, -50])
  * //= [ 87.5, -47.5 ]
  */
 export function bboxToCenter (bbox) {
-  const [west, south, east, north] = bbox
-  let lng = (west - east) / 2 + east
-  let lat = (south - north) / 2 + north
+  var west = bbox[0]
+  var south = bbox[1]
+  var east = bbox[2]
+  var north = bbox[3]
+  var lng = (west - east) / 2 + east
+  var lat = (south - north) / 2 + north
   lng = Number(lng.toFixed(6))
   lat = Number(lat.toFixed(6))
   return [lng, lat]
@@ -39,13 +44,15 @@ export function bboxToCenter (bbox) {
  * @param {LngLat} lnglat Longitude (Meridians) & Latitude (Parallels) in decimal degrees
  * @returns {Meters} Meters coordinates
  * @example
- * const meters = lngLatToMeters([126, 37])
+ * var meters = lngLatToMeters([126, 37])
  * //=[ 14026255.8, 4439106.7 ]
  */
 export function lngLatToMeters (lnglat) {
-  const [lng, lat] = validateLngLat(lnglat)
-  let x = lng * originShift / 180.0
-  let y = Math.log(Math.tan((90 + lat) * Math.PI / 360.0)) / (Math.PI / 180.0)
+  validateLngLat(lnglat)
+  var lng = lnglat[0]
+  var lat = lnglat[1]
+  var x = lng * originShift / 180.0
+  var y = Math.log(Math.tan((90 + lat) * Math.PI / 360.0)) / (Math.PI / 180.0)
   y = y * originShift / 180.0
   x = Number(x.toFixed(1))
   y = Number(y.toFixed(1))
@@ -58,13 +65,14 @@ export function lngLatToMeters (lnglat) {
  * @param {Meters} meters Meters in Mercator [x, y]
  * @returns {LngLat} LngLat coordinates
  * @example
- * const lnglat = metersToLngLat([14026255, 4439106])
+ * var lnglat = metersToLngLat([14026255, 4439106])
  * //=[ 126, 37 ]
  */
 export function metersToLngLat (meters) {
-  const [x, y] = meters
-  let lng = (x / originShift) * 180.0
-  let lat = (y / originShift) * 180.0
+  var x = meters[0]
+  var y = meters[1]
+  var lng = (x / originShift) * 180.0
+  var lat = (y / originShift) * 180.0
   lat = 180 / Math.PI * (2 * Math.atan(Math.exp(lat * Math.PI / 180.0)) - Math.PI / 2.0)
   lng = Number(lng.toFixed(6))
   lat = Number(lat.toFixed(6))
@@ -79,14 +87,15 @@ export function metersToLngLat (meters) {
  * @param {number} [tileSize=256] Tile size
  * @returns {Pixels} Pixels coordinates
  * @example
- * const pixels = metersToPixels([14026255, 4439106], 13)
+ * var pixels = metersToPixels([14026255, 4439106], 13)
  * //=[ 1782579.1, 1280877.3, 13 ]
  */
 export function metersToPixels (meters, zoom, tileSize) {
-  const [x, y] = meters
-  const res = resolution(zoom, tileSize)
-  const px = (x + originShift) / res
-  const py = (y + originShift) / res
+  var x = meters[0]
+  var y = meters[1]
+  var res = resolution(zoom, tileSize)
+  var px = (x + originShift) / res
+  var py = (y + originShift) / res
   return [px, py, zoom]
 }
 
@@ -97,12 +106,12 @@ export function metersToPixels (meters, zoom, tileSize) {
  * @param {number} zoom Zoom level
  * @returns {Tile} TMS Tile
  * @example
- * const tile = lngLatToTile([126, 37], 13)
+ * var tile = lngLatToTile([126, 37], 13)
  * //=[ 6963, 5003, 13 ]
  */
 export function lngLatToTile (lnglat, zoom) {
-  const meters = lngLatToMeters(validateLngLat(lnglat))
-  const pixels = metersToPixels(meters, zoom)
+  var meters = lngLatToMeters(validateLngLat(lnglat))
+  var pixels = metersToPixels(meters, zoom)
   return pixelsToTile(pixels)
 }
 
@@ -113,14 +122,14 @@ export function lngLatToTile (lnglat, zoom) {
  * @param {number} zoom Zoom level
  * @returns {Google} Google (XYZ) Tile
  * @example
- * const google = lngLatToGoogle([126, 37], 13)
+ * var google = lngLatToGoogle([126, 37], 13)
  * //=[ 6963, 3188, 13 ]
  */
 export function lngLatToGoogle (lnglat, zoom) {
   if (zoom === 0) {
     return [0, 0, 0]
   }
-  const tile = lngLatToTile(validateLngLat(lnglat), zoom)
+  var tile = lngLatToTile(validateLngLat(lnglat), zoom)
   return tileToGoogle(tile)
 }
 
@@ -131,14 +140,14 @@ export function lngLatToGoogle (lnglat, zoom) {
  * @param {number} zoom Zoom level
  * @returns {Tile} TMS Tile
  * @example
- * const tile = metersToTile([14026255, 4439106], 13)
+ * var tile = metersToTile([14026255, 4439106], 13)
  * //=[ 6963, 5003, 13 ]
  */
 export function metersToTile (meters, zoom) {
   if (zoom === 0) {
     return [0, 0, 0]
   }
-  const pixels = metersToPixels(meters, zoom)
+  var pixels = metersToPixels(meters, zoom)
   return pixelsToTile(pixels)
 }
 
@@ -149,14 +158,17 @@ export function metersToTile (meters, zoom) {
  * @param {number} [tileSize=256] Tile size
  * @returns {Meters} Meters coordinates
  * @example
- * const meters = pixelsToMeters([1782579, 1280877, 13])
+ * var meters = pixelsToMeters([1782579, 1280877, 13])
  * //=[ 14026252.0, 4439099.5 ]
  */
 export function pixelsToMeters (pixels, tileSize) {
-  const [px, py, zoom] = validatePixels(pixels)
-  const res = resolution(zoom, tileSize)
-  let mx = px * res - originShift
-  let my = py * res - originShift
+  validatePixels(pixels)
+  var px = pixels[0]
+  var py = pixels[1]
+  var zoom = pixels[2]
+  var res = resolution(zoom, tileSize)
+  var mx = px * res - originShift
+  var my = py * res - originShift
   mx = Number(mx.toFixed(1))
   my = Number(my.toFixed(1))
   return [mx, my]
@@ -169,16 +181,20 @@ export function pixelsToMeters (pixels, tileSize) {
  * @param {number} [tileSize=256] Tile size
  * @returns {Tile} TMS Tile
  * @example
- * const tile = pixelsToTile([1782579, 1280877, 13])
+ * var tile = pixelsToTile([1782579, 1280877, 13])
  * //=[ 6963, 5003, 13 ]
  */
-export function pixelsToTile (pixels, tileSize = 256) {
-  const [px, py, zoom] = validatePixels(pixels)
+export function pixelsToTile (pixels, tileSize) {
+  tileSize = tileSize || 256
+  validatePixels(pixels)
+  var px = pixels[0]
+  var py = pixels[1]
+  var zoom = pixels[2]
   if (zoom === 0) {
     return [0, 0, 0]
   }
-  let tx = Math.ceil(px / tileSize) - 1
-  let ty = Math.ceil(py / tileSize) - 1
+  var tx = Math.ceil(px / tileSize) - 1
+  var ty = Math.ceil(py / tileSize) - 1
   if (tx < 0) {
     tx = 0
   }
@@ -198,13 +214,17 @@ export function pixelsToTile (pixels, tileSize = 256) {
  * @param {number} [tileSize=256] Tile size
  * @returns {BBox} bbox extent in [minX, minY, maxX, maxY] order
  * @example
- * const bbox = tileToBBoxMeters([6963, 5003, 13])
+ * var bbox = tileToBBoxMeters([6963, 5003, 13])
  * //=[ 14025277.4, 4437016.6, 14030169.4, 4441908.5 ]
  */
-export function tileToBBoxMeters (tile, tileSize = 256) {
-  const [tx, ty, zoom] = validateTile(tile)
-  let min = pixelsToMeters([tx * tileSize, ty * tileSize, zoom])
-  let max = pixelsToMeters([(tx + 1) * tileSize, (ty + 1) * tileSize, zoom])
+export function tileToBBoxMeters (tile, tileSize) {
+  tileSize = tileSize || 256
+  validateTile(tile)
+  var tx = tile[0]
+  var ty = tile[1]
+  var zoom = tile[2]
+  var min = pixelsToMeters([tx * tileSize, ty * tileSize, zoom])
+  var max = pixelsToMeters([(tx + 1) * tileSize, (ty + 1) * tileSize, zoom])
   return [min[0], min[1], max[0], max[1]]
 }
 
@@ -217,17 +237,24 @@ export function tileToBBoxMeters (tile, tileSize = 256) {
  * @param {number} zoom Zoom level
  * @returns {BBox} bbox extent in [minX, minY, maxX, maxY] order
  * @example
- * const bbox = tileToBBox([6963, 5003, 13])
+ * var bbox = tileToBBox([6963, 5003, 13])
  * //=[ 125.991, 36.985, 126.035, 37.020 ]
  */
 export function tileToBBox (tile) {
-  const [tx, ty, zoom] = validateTile(tile)
+  validateTile(tile)
+  var tx = tile[0]
+  var ty = tile[1]
+  var zoom = tile[2]
   if (zoom === 0) {
     return [-180, -85.051129, 180, 85.051129]
   }
-  const [mx1, my1, mx2, my2] = tileToBBoxMeters([tx, ty, zoom])
-  const min = metersToLngLat([mx1, my1, zoom])
-  const max = metersToLngLat([mx2, my2, zoom])
+  var bbox = tileToBBoxMeters([tx, ty, zoom])
+  var mx1 = bbox[0]
+  var my1 = bbox[1]
+  var mx2 = bbox[2]
+  var my2 = bbox[3]
+  var min = metersToLngLat([mx1, my1, zoom])
+  var max = metersToLngLat([mx2, my2, zoom])
   return [min[0], min[1], max[0], max[1]]
 }
 
@@ -237,11 +264,11 @@ export function tileToBBox (tile) {
  * @param {Google} google Google [x, y, zoom]
  * @returns {BBox} bbox extent in [minX, minY, maxX, maxY] order
  * @example
- * const bbox = googleToBBoxMeters([6963, 3188, 13])
+ * var bbox = googleToBBoxMeters([6963, 3188, 13])
  * //=[ 14025277.4, 4437016.6, 14030169.4, 4441908.5 ]
  */
 export function googleToBBoxMeters (google) {
-  const Tile = googleToTile(google)
+  var Tile = googleToTile(google)
   return tileToBBoxMeters(Tile)
 }
 
@@ -251,11 +278,11 @@ export function googleToBBoxMeters (google) {
  * @param {Google} google Google [x, y, zoom]
  * @returns {BBox} bbox extent in [minX, minY, maxX, maxY] order
  * @example
- * const bbox = googleToBBox([6963, 3188, 13])
+ * var bbox = googleToBBox([6963, 3188, 13])
  * //=[ 125.991, 36.985, 126.035, 37.020 ]
  */
 export function googleToBBox (google) {
-  const Tile = googleToTile(google)
+  var Tile = googleToTile(google)
   return tileToBBox(Tile)
 }
 
@@ -265,16 +292,19 @@ export function googleToBBox (google) {
  * @param {Tile} tile Tile [x, y, zoom]
  * @returns {Google} Google (XYZ) Tile
  * @example
- * const google = tileToGoogle([6963, 5003, 13])
+ * var google = tileToGoogle([6963, 5003, 13])
  * //=[ 6963, 3188, 13 ]
  */
 export function tileToGoogle (tile) {
-  const [tx, ty, zoom] = validateTile(tile)
+  validateTile(tile)
+  var tx = tile[0]
+  var ty = tile[1]
+  var zoom = tile[2]
   if (zoom === 0) {
     return [0, 0, 0]
   }
-  const x = tx
-  const y = (Math.pow(2, zoom) - 1) - ty
+  var x = tx
+  var y = (Math.pow(2, zoom) - 1) - ty
   return [x, y, zoom]
 }
 
@@ -284,13 +314,15 @@ export function tileToGoogle (tile) {
  * @param {Google} google Google [x, y, zoom]
  * @returns {Tile} TMS Tile
  * @example
- * const tile = googleToTile([6963, 3188, 13])
+ * var tile = googleToTile([6963, 3188, 13])
  * //=[ 6963, 5003, 13 ]
  */
 export function googleToTile (google) {
-  const [x, y, zoom] = google
-  const tx = x
-  const ty = Math.pow(2, zoom) - y - 1
+  var x = google[0]
+  var y = google[1]
+  var zoom = google[2]
+  var tx = x
+  var ty = Math.pow(2, zoom) - y - 1
   return [tx, ty, zoom]
 }
 
@@ -300,11 +332,11 @@ export function googleToTile (google) {
  * @param {Google} google Google [x, y, zoom]
  * @returns {string} Microsoft's Quadkey schema
  * @example
- * const quadkey = googleToQuadkey([6963, 3188, 13])
+ * var quadkey = googleToQuadkey([6963, 3188, 13])
  * //='1321102330211'
  */
 export function googleToQuadkey (google) {
-  const Tile = googleToTile(google)
+  var Tile = googleToTile(google)
   return tileToQuadkey(Tile)
 }
 
@@ -314,20 +346,23 @@ export function googleToQuadkey (google) {
  * @param {Tile} tile Tile [x, y, zoom]
  * @returns {string} Microsoft's Quadkey schema
  * @example
- * const quadkey = tileToQuadkey([6963, 5003, 13])
+ * var quadkey = tileToQuadkey([6963, 5003, 13])
  * //='1321102330211'
  */
 export function tileToQuadkey (tile) {
-  let [tx, ty, zoom] = validateTile(tile)
+  validateTile(tile)
+  var tx = tile[0]
+  var ty = tile[1]
+  var zoom = tile[2]
   // Zoom 0 does not exist for Quadkey
   if (zoom === 0) {
     return ''
   }
-  let quadkey = ''
+  var quadkey = ''
   ty = (Math.pow(2, zoom) - 1) - ty
-  range(zoom, 0, -1).map(i => {
-    let digit = 0
-    let mask = 1 << (i - 1)
+  range(zoom, 0, -1).map(function (i) {
+    var digit = 0
+    var mask = 1 << (i - 1)
     if ((tx & mask) !== 0) {
       digit += 1
     }
@@ -345,11 +380,11 @@ export function tileToQuadkey (tile) {
  * @param {string} quadkey Microsoft's Quadkey schema
  * @returns {Tile} TMS Tile
  * @example
- * const tile = quadkeyToTile('1321102330211')
+ * var tile = quadkeyToTile('1321102330211')
  * //=[ 6963, 5003, 13 ]
  */
 export function quadkeyToTile (quadkey) {
-  const Google = quadkeyToGoogle(quadkey)
+  var Google = quadkeyToGoogle(quadkey)
   return googleToTile(Google)
 }
 
@@ -359,15 +394,15 @@ export function quadkeyToTile (quadkey) {
  * @param {string} quadkey Microsoft's Quadkey schema
  * @returns {Google} Google (XYZ) Tile
  * @example
- * const google = quadkeyToGoogle('1321102330211')
+ * var google = quadkeyToGoogle('1321102330211')
  * //=[ 6963, 3188, 13 ]
  */
 export function quadkeyToGoogle (quadkey) {
-  let x = 0
-  let y = 0
-  const zoom = quadkey.length
-  range(zoom, 0, -1).map(i => {
-    let mask = 1 << (i - 1)
+  var x = 0
+  var y = 0
+  var zoom = quadkey.length
+  range(zoom, 0, -1).map(function (i) {
+    var mask = 1 << (i - 1)
     switch (parseInt(quadkey[zoom - i], 0)) {
       case 0:
         break
@@ -394,12 +429,12 @@ export function quadkeyToGoogle (quadkey) {
  * @param {BBox} bbox extent in [minX, minY, maxX, maxY] order
  * @returns {BBox} bbox extent in [minX, minY, maxX, maxY] order
  * @example
- * const meters = bboxToMeters([ 125, 35, 127, 37 ])
+ * var meters = bboxToMeters([ 125, 35, 127, 37 ])
  * //=[ 13914936.3, 4163881.1, 14137575.3, 4439106.7 ]
  */
 export function bboxToMeters (bbox) {
-  const min = lngLatToMeters([bbox[0], bbox[1]])
-  const max = lngLatToMeters([bbox[2], bbox[3]])
+  var min = lngLatToMeters([bbox[0], bbox[1]])
+  var max = lngLatToMeters([bbox[2], bbox[3]])
   return [min[0], min[1], max[0], max[1]]
 }
 
@@ -418,13 +453,15 @@ export function bboxToMeters (bbox) {
  * //= Error: Illegal parameters for tile
  */
 export function validateTile (tile) {
-  const [tx, ty, zoom] = tile
+  var tx = tile[0]
+  var ty = tile[1]
+  var zoom = tile[2]
   validateZoom(zoom)
   if (tx === undefined || tx === null) { throw new Error('<x> is required') }
   if (ty === undefined || ty === null) { throw new Error('<y> is required') }
   if (tx < 0) { throw new Error('<x> must not be less than 0') }
   if (ty < 0) { throw new Error('<y> must not be less than 0') }
-  const maxCount = Math.pow(2, zoom)
+  var maxCount = Math.pow(2, zoom)
   if (tx >= maxCount || ty >= maxCount) { throw new Error('Illegal parameters for tile') }
   return tile
 }
@@ -463,7 +500,8 @@ export function validateZoom (zoom) {
  * //= Error: LngLat [lng] must be within -180 to 180 degrees
  */
 export function validateLngLat (lnglat) {
-  const [lng, lat] = lnglat
+  var lng = lnglat[0]
+  var lat = lnglat[1]
   if (lat === undefined || lat === null) { throw new Error('<lat> is required') }
   if (lng === undefined || lng === null) { throw new Error('<lng> is required') }
   if (lat < -90 || lat > 90) { throw new Error('LngLat <lat> must be within -90 to 90 degrees') }
@@ -496,7 +534,7 @@ export function validatePixels (pixels) {
  * @param {number} [tileSize=256] Tile size
  * @returns {number} resolution
  * @example
- * const res = resolution(13)
+ * var res = resolution(13)
  * //=19.109257071294063
  */
 export function resolution (zoom, tileSize) {
@@ -527,9 +565,9 @@ export function range (start, stop, step) {
   if (!step) {
     step = stop < start ? -1 : 1
   }
-  const length = Math.max(Math.ceil((stop - start) / step), 0)
-  const range = Array(length)
-  for (let idx = 0; idx < length; idx++, start += step) {
+  var length = Math.max(Math.ceil((stop - start) / step), 0)
+  var range = Array(length)
+  for (var idx = 0; idx < length; idx++, start += step) {
     range[idx] = start
   }
   return range